refactor(moyo): gather layer base modules into base/layer
Move `lattice_2d.rs`, `layer_lattice.rs`, and `layer_cell.rs` from
`moyo/src/base/` into a new `moyo/src/base/layer/` submodule. The new
`moyo/src/base/layer.rs` declares the three private files and re-exports
`Lattice2D`, `LayerLattice`, `LayerCell`, so the public API surface
(`crate::base::{Lattice2D, LayerCell, LayerLattice}`) is unchanged.

Internal `super::*` paths in the moved files become `super::super::*`
for sibling base modules (`error`, `lattice`, `cell`, `tolerance`); the
intra-`layer/` import (`layer_cell` -> `layer_lattice`) stays as
`super::layer_lattice`. `base/transformation.rs` updated to use
`super::layer::{LayerCell, LayerLattice}`.

No behavior change; all 178 cargo tests pass.
